+++ tests/TestCase.php
@@ -11,7 +11,6 @@ use Filament\FilamentServiceProvider;
 use Filament\Forms\FormsServiceProvider;
 use Filament\Infolists\InfolistsServiceProvider;
 use Filament\Notifications\NotificationsServiceProvider;
-use Filament\Schemas\SchemasServiceProvider;
 use Filament\Support\Livewire\Partials\DataStoreOverride;
 use Filament\Support\SupportServiceProvider;
 use Filament\Tables\TablesServiceProvider;
@@ -59,7 +58,6 @@ abstract class TestCase extends BaseTestCase
             SupportServiceProvider::class,
             FilamentServiceProvider::class,
             FormsServiceProvider::class,
-            SchemasServiceProvider::class,
             TablesServiceProvider::class,
             ActionsServiceProvider::class,
             InfolistsServiceProvider::class,